Sun, Mar 6, 2022, 12:38 PM +00:00

+++ readability/readability.c
@@ -11,14 +11,14 @@ int main(void)
     string txt = get_string("Text: ");
 
     int letters = count_letters(txt);
-    float words = count_words(txt);
-    float sentences = count_sentences(txt);
+    int words = count_words(txt);
+    int sentences = count_sentences(txt);
     float L = letters / words * 100;
     float S = sentences / words * 100;
     int index = 0.0588 * L - 0.296 * S - 15.8;
     printf("%i letters\n", letters);
-    printf("%f words\n", words + 1);
-    printf("%f sentences\n", sentences);
+    printf("%i words\n", words + 1);
+    printf("%i sentences\n", sentences);
     printf("%i index\n", index);
 }
